services/users/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
SQLAlchemy==2.0.32
passlib[bcrypt]==1.7.4
PyJWT==2.9.0


